refactor(Codecs): Move Codec interface to own file
