+++ src/codecs/dar/index.ts
@@ -22,7 +22,8 @@ import produce from 'immer'
 import { html as beautifyHtml } from 'js-beautify'
 import path from 'path'
 import tempy from 'tempy'
-import { Encode, EncodeOptions, write } from '../..'
+import { EncodeOptions, write } from '../..'
+import { Encode } from '../types'
 import * as uri from '../../util/uri'
 import * as vfile from '../../util/vfile'
