+++ src/codecs/html/index.ts
@@ -16,7 +16,8 @@ import { html as beautifyHtml } from 'js-beautify'
 import jsdom from 'jsdom'
 import JSON5 from 'json5'
 import path from 'path'
-import { Encode, EncodeOptions } from '../..'
+import { EncodeOptions } from '../..'
+import { Encode } from '../types'
 import { columnIndexToName } from '../../codecs/xlsx'
 import bundle from '../../util/bundle'
 import * as vfile from '../../util/vfile'
