+++ src/codecs/jats-pandoc/index.ts
@@ -3,7 +3,7 @@
  */
 
 import stencila from '@stencila/schema'
-import { Encode } from '../..'
+import { Encode } from '../types'
 import * as pandoc from '../pandoc'
 import * as vfile from '../../util/vfile'
