+++ src/codecs/rpng/index.ts
@@ -9,10 +9,11 @@ import pngText from 'png-chunk-text'
 import pngEncode from 'png-chunks-encode'
 import pngExtract, { Chunk } from 'png-chunks-extract'
 import punycode from 'punycode/'
-import { dump, Encode, EncodeOptions } from '../../index'
-import * as puppeteer from '../../util/puppeteer'
+import { dump, EncodeOptions } from '../../index'
 import bundle from '../../util/bundle'
+import * as puppeteer from '../../util/puppeteer'
 import * as vfile from '../../util/vfile'
+import { Encode } from '../types'
 
 // A vendor media type similar to https://www.iana.org/assignments/media-types/image/vnd.mozilla.apng
 // an custom extension to be able to refere to this format more easily.
